listing all items api is added and its working fine
