+++ backend/app.js
@@ -33,7 +33,7 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
-app.use('/api',apirouter);
+app.use('/api',apiRouter);
 // app.use('/', indexRouter);
 // app.use('/users', usersRouter);
